// src/utils/helper.d.ts
export declare const asyncHandler: (fn: Function) => (req: any, res: any, next: any) => void;
export declare const generateRandomColor: () => string;
export declare const exclude: <T extends Record<string, any>, K extends keyof T>(obj: T, keys: K[]) => Omit<T, K>;
//# sourceMappingURL=helper.d.ts.map